Detects and fixes changed <info> (fixes #5), added detection fatal error (fixes #10) and fixes reading of <info> part.
Gmailcc now correctly detects and fixes flags when a local client
changed them on one or more links of an email.
We encountered a fatal error in connection during testing so we added
handling for it.
Apparently flags from the <info> part of the filename weren't read
correctly which caused a lot of flag-changing. We fixed this but
ultimately it won't matter much since we're now synching the flags if
they are changed or not, to fix #5.